% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-engin.R
\name{feature_engin}
\alias{feature_engin}
\title{Feature Engineering}
\usage{
feature_engin(nomInd, annee)
}
\arguments{
\item{nomInd}{Le nom de l'indice}

\item{annee}{l'année}
}
\value{
Générer les nouvelles variables pour une période donnée
}
\description{
Feature Engineering
}
\examples{
resultat_DJI <- feature_engin(DJI,'2021::')
View(resultat_DJI)
}
