% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele logistique.R
\name{glm_fit_indice}
\alias{glm_fit_indice}
\title{modele logistique}
\usage{
glm_fit_indice(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Nom de l'indice version data frame}
}
\value{
Retourner la variable glm_fit de l'indice
}
\description{
modele logistique
}
\examples{
glm_DJI <- glm_fits_indice(df_DJI)
summary(glm_DJI)

glm.probs=predict(glm_DJI,type="response")
glm.probs[1:5]
glm.pred=ifelse(glm.probs>0.5,"Up","Down")
glm.pred[1:5]

attach(df_DJI)
table(glm.pred,direction)

}
