% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele linéaire high-close.R
\name{modele_lineaire_highclose}
\alias{modele_lineaire_highclose}
\title{Modélisation linéaire high close}
\usage{
modele_lineaire_highclose(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Le nom de l"indice version data frame}
}
\value{
le modele linéaire de l'indice qui prédit la variable Close dépéndant de la variable High
}
\description{
Modélisation linéaire high close
}
