% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele linéaire low-close.R
\name{modele_lineaire_lowclose}
\alias{modele_lineaire_lowclose}
\title{Modélisation linéaire low close}
\usage{
modele_lineaire_lowclose(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Le nom de l"indice version data frame}
}
\value{
le modele linéaire de l'indice qui prédit la variable Close dépéndant de la variable Low
}
\description{
Modélisation linéaire low close
}
