% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele linéaire open-close.R
\name{modele_lineaire_openclose}
\alias{modele_lineaire_openclose}
\title{Modélisation linéaire open close}
\usage{
modele_lineaire_openclose(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Le nom de l"indice version data frame}
}
\value{
le modele linéaire de l'indice qui prédit la variable Close dépéndant de la variable Open
}
\description{
Modélisation linéaire open close
}
