% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{create a barplot}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of bernoulli trials}

\item{p}{probability of success}
}
\value{
it return a barplot of the proportions and table of relative frequencies
}
\description{
create a barplot
}
\examples{
mybin(100,10,0.5)
}
