% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab10.R
\name{myci}
\alias{myci}
\title{95% confidence interval}
\usage{
myci(x)
}
\arguments{
\item{x}{sample}
}
\value{
95% confidence interval for U from a single sample x
}
\description{
95% confidence interval
}
\examples{
set.seed(23); x = rnorm(30,mean=10,sd=12)
myci(x)
}
