% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8.R
\name{myclt}
\alias{myclt}
\title{histogram of sample means}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{number of binomial distribution}

\item{iter}{number of iterations}

\item{a}{min}

\item{b}{max}
}
\value{
histogram of sample means
}
\description{
histogram of sample means
}
\examples{

w=myclt(n=10,iter=10000)
}
