% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{myf}
\alias{myf}
\title{Find the Standard Deviation of x}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
Standard Deviation of x
}
\description{
Find the Standard Deviation of x
}
\examples{
x <- c(179,160,136,227)
sd(x)

}
