% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{Remove columns with all 0s in the matrix}
\usage{
rm_0s_by_col(matrix_data)
}
\arguments{
\item{matrix_data}{Numeric matrix}
}
\value{
Matrix with columns having all 0s removed
}
\description{
Remove columns with all 0s in the matrix
}
