% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{Remove rows with all 0s in the matrix}
\usage{
rm_0s_by_row(matrix_data)
}
\arguments{
\item{matrix_data}{Numeric matrix}
}
\value{
Matrix with rows having all 0s removed
}
\description{
Remove rows with all 0s in the matrix
}
